Moved to StudentController folder
